requirements.txt
streamlit
langchain
langchain-community
transformers
torch
torchvision
pillow
accelerate
sentencepiece